parent info

+++ data/migrations/20190926131952_school.js
@@ -40,10 +40,10 @@ exports.up = function (knex, Promise) {
         .createTable('parents', tbl => {
             tbl.increments();
             tbl.string('name', 255).notNullable();
-            tbl.integer('phone', 50).notNullable();
+            tbl.bigInteger('phone', 50).notNullable();
             tbl.string('address', 25);
             tbl.string('spouse_name', 25);
-            tbl.integer('spouse_phone', 25);
+            tbl.bigInteger('spouse_phone', 25);
 
         })
